extract mocks and stubs
