+++ Tests/NotifyTests/Mocks/MockNotifyUpdateRequester.swift
@@ -1,0 +1,15 @@
+
+import Foundation
+@testable import WalletConnectPush
+
+
+class MockNotifyUpdateRequester: NotifyUpdateRequesting {
+    var updatedTopics: [String] = []
+    var completionHandler: (() -> Void)?
+
+    func update(topic: String, scope: Set<String>) async throws {
+        updatedTopics.append(topic)
+        completionHandler?()
+    }
+}
+
