+++ Tests/NotifyTests/SubscriptionsAutoUpdaterTests.swift
@@ -34,64 +34,3 @@ class SubscriptionsAutoUpdaterTests: XCTestCase {
         XCTAssertEqual(notifyUpdateRequester.updatedTopics, ["topic2"])
     }
 }
-
-
-extension PushSubscription {
-    static func stub(topic: String, expiry: Date) -> PushSubscription {
-        let account = Account(chainIdentifier: "eip155:1", address: "0x15bca56b6e2728aec2532df9d436bd1600e86688")!
-        let relay = RelayProtocolOptions.stub()
-        let metadata = AppMetadata.stub()
-        let symKey = "key1"
-
-        return PushSubscription(
-            topic: topic,
-            account: account,
-            relay: relay,
-            metadata: metadata,
-            scope: ["test": ScopeValue(description: "desc", enabled: true)],
-            expiry: expiry,
-            symKey: symKey
-        )
-    }
-}
-
-
-class MockNotifyUpdateRequester: NotifyUpdateRequesting {
-    var updatedTopics: [String] = []
-    var completionHandler: (() -> Void)?
-
-    func update(topic: String, scope: Set<String>) async throws {
-        updatedTopics.append(topic)
-        completionHandler?()
-    }
-}
-
-
-
-class MockPushStoring: PushStoring {
-    var subscriptions: [PushSubscription]
-
-    init(subscriptions: [PushSubscription]) {
-        self.subscriptions = subscriptions
-    }
-
-    func getSubscriptions() -> [PushSubscription] {
-        return subscriptions
-    }
-
-    func getSubscription(topic: String) -> PushSubscription? {
-        return subscriptions.first { $0.topic == topic }
-    }
-
-    func setSubscription(_ subscription: PushSubscription) async throws {
-        if let index = subscriptions.firstIndex(where: { $0.topic == subscription.topic }) {
-            subscriptions[index] = subscription
-        } else {
-            subscriptions.append(subscription)
-        }
-    }
-
-    func deleteSubscription(topic: String) async throws {
-        subscriptions.removeAll(where: { $0.topic == topic })
-    }
-}
